# tacs.api/src/main/resources/swagger/api.yaml
openapi: 3.0.1
info:
  title: TP TACS
  version: 1.0.0
servers:
  - url: http://127.0.0.1:8080/api/v1
...
paths:
...
  /usuarios:
    get:
      tags:
        - Usuarios
      summary: Permite listar a los usuarios
      operationId: listarUsuarios
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: filter
          schema:
            type: string
          required: false
        - in: query
          name: tamanioPagina
          schema:
            type: integer
            format: int64
            default: 10
          required: false
        - in: query
          name: pagina
          schema:
            type: integer
            format: int64
            default: 0
          required: false
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                title: listarUsuariosResponse
                properties:
                  usuarios:
                    type: array
                    items:
                      $ref: '#/components/schemas/UsuarioModel'
        400:
          description: Bad request
        403:
          description: Forbidden
        404:
          description: Page not found
...
components:
  schemas:
...
    UsuarioModel:
      type: object
      description: Un usuario.
      required:
        - id
        - nombreDeUsuario
      properties:
        id:
          type: integer
          format: int64
        nombreDeUsuario:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT